refactor: use single base class
