+++ haystack_experimental/core/super_component/__init__.py
@@ -2,7 +2,6 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from .base import SuperComponentBase
 from .super_component import SuperComponent
 
-_all_ = ["SuperComponent", "SuperComponentBase"]
+_all_ = ["SuperComponent"]
